+++ lib/prefer_sdk/io.dart
@@ -12,16 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Cross-platform implementation of 'dart:io'.
-///
-/// To deal with limitations of conditional imports/exports in Dart, developers
-/// can choose from three slightly libraries:
-///   * 'package:universal_io/prefer_sdk/io.dart'
-///     * Exports SDK version by default
-///   * 'package:universal_io/prefer_universal/io.dart'
-///     * Exports non-SDK version by default
-///   * 'package:universal_io/io.dart'
-///     * Exports either one of the above libraries.
-library universal_io;
-
-export 'prefer_sdk/io.dart';
+export '../src/sdk_io.dart' if (dart.library.html) '../src/io/io.dart';