+++ lib/src/driver_base/base_raw_socket.dart
@@ -17,7 +17,7 @@ import 'dart:typed_data';
 
 import 'package:meta/meta.dart';
 import 'package:universal_io/driver_base.dart' show Uint8ListBuffer;
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 abstract class BaseRawSocket extends Stream<RawSocketEvent>
     implements RawSocket {