+++ lib/src/driver_for_browser/browser_http_client.dart
@@ -16,7 +16,7 @@ import 'dart:async';
 import 'dart:html' as html;
 
 import 'package:universal_io/driver_base.dart';
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 import 'browser_http_client_request.dart';