+++ lib/src/driver_for_browser/browser_http_client_exception.dart
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import 'package:meta/meta.dart';
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 import 'browser_http_client.dart';