+++ lib/src/driver_for_browser/browser_http_client_request.dart
@@ -18,7 +18,7 @@ import 'dart:typed_data';
 
 import 'package:universal_io/driver_base.dart';
 import 'package:universal_io/driver_base.dart' show Uint8ListBuffer;
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 import 'browser_http_client.dart';
 import 'browser_http_client_exception.dart';