+++ test/src/test_suite/http_client.dart
@@ -17,7 +17,7 @@ import 'dart:convert';
 
 import 'package:meta/meta.dart';
 import 'package:test/test.dart';
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 import 'example_http_server.dart';