+++ test/src/test_suite/internet_address.dart
@@ -16,7 +16,7 @@ import 'dart:typed_data';
 
 import 'package:test/test.dart';
 import 'package:universal_io/driver.dart';
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 void testInternetAddress() {
   group("InternetAddress", () {