+++ test/src/test_suite/process.dart
@@ -18,7 +18,7 @@ library process_test;
 import 'dart:convert';
 
 import 'package:test/test.dart';
-import 'package:universal_io/io.dart';
+import 'package:universal_io/prefer_universal/io.dart';
 
 void testProcess({bool isPosix = true}) {
   group("Process", () {